% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_fun.R
\name{ols_fun}
\alias{ols_fun}
\title{OLS Log-likelihood}
\usage{
ols_fun(dat, param)
}
\arguments{
\item{dat}{Dataframe to be used}

\item{param}{Vector of parameters to be inserted}
}
\value{
The Log-likelihood of the OLS
}
\description{
\code{ols_fun} Gives the OLS likelihood given the parameters
}
\examples{
x <- runif(100, 0, 10)
y <- 2*x + rnorm(100,0,1)
dat <- data.frame(y,x)
ols_fun(dat, c(2,1))
}
