% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihoods.R
\name{OlsFun}
\alias{OlsFun}
\title{OLS Log-likelihood}
\usage{
OlsFun(param)
}
\arguments{
\item{param}{Vector of parameters to be inserted}
}
\value{
The Log-likelihood
}
\description{
Calculates the OLS likelihood given the parameters
}

