% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatFrontier.R
\docType{package}
\name{SpatFrontier}
\alias{SpatFrontier}
\alias{SpatFrontier-package}
\title{SpatFrontier: a package for the estimation of spatial frontier models}
\description{
The SpatFrontier package provides at the moment one important function:
spatial_frontier.
}
\section{SpatFrontier functions}{

The SpatFrontier functions
}

