% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_error_frontier_fun.R
\name{spatial_error_frontier_fun}
\alias{spatial_error_frontier_fun}
\title{Spatial error frontier distribution}
\usage{
spatial_error_frontier_fun(y, X, param, W)
}
\arguments{
\item{y}{The endogeneous variable}

\item{X}{The dataframe}

\item{param}{The parameters to be estimated}

\item{W}{The spatial weight matrix}
}
\description{
Spatial error frontier distribution
}
