% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_frontier.R
\name{spatial_frontier}
\alias{spatial_frontier}
\title{spatial_frontier}
\usage{
spatial_frontier(formula, data = NULL, data_w = NULL, frontier = TRUE)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Dataframe to be used}

\item{data_w}{Spatial weights matrix}

\item{frontier}{if TRUE a frontier analysis is executed}
}
\value{
list with fitted elements from a maximum likelihood estimation
}
\description{
spatial_frontier
}
